wire up types correctly
